+++ src/main/java/wooteco/chess/domain/repository/BoardRepository.java
@@ -1,13 +1,15 @@
-package wooteco.chess.domain.board;
+package wooteco.chess.domain.repository;
 
 import org.springframework.data.jdbc.repository.query.Modifying;
 import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
+@Component
 @Repository
 public interface BoardRepository extends CrudRepository<BoardEntity, Long> {
 
@@ -23,7 +25,5 @@ public interface BoardRepository extends CrudRepository<BoardEntity, Long> {
     @Modifying
     @Query("DELETE FROM board WHERE roomId = :roomId")
     void deleteByRoomId(@Param("roomId") Long roomId);
-
-
 
 }
