+++ src/main/resources/templates/index.hbs
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="ko">
+<html lang="en">
 <head>
     <meta charset="UTF-8">
     <title>체스게임 by BossDog & Orange</title>
