+++ src/main/resources/templates/room.hbs
@@ -18,7 +18,6 @@
     </header>
 
 
-
     <!-- CONTROLLER -->
     <div class="button-controller">
         <form action="/room/{{roomId}}/reset" method="get">
